refactor: removed the graphwithoutdashboard component
